models v1.2

+++ Pharmacy/models.py
@@ -15,7 +15,7 @@ class Drug(models.Model):
     description = models.CharField(max_length=100, null=True)
     stock_level = models.PositiveIntegerField(default=0)
     expiry_date = models.DateField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=15, decimal_places=2)
 
     def __str__(self):
         return self.name + ' ' + self.brand_name
